Added Exception Handling and further comments

+++ src/main/java/com/dockerators/bookapp/service/BookServiceImpl.java
@@ -2,6 +2,9 @@ package com.dockerators.bookapp.service;
 
 import com.dockerators.bookapp.dao.BookRepository;
 import com.dockerators.bookapp.entity.Book;
+import com.dockerators.bookapp.exception.BookAlreadyExistsException;
+import com.dockerators.bookapp.exception.BookNotFoundException;
+import com.dockerators.bookapp.exception.NullFieldsException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -30,26 +33,63 @@ public class BookServiceImpl implements BookService {
     public Book findById(int id) {
         // May be null.
         Optional<Book> result = this.bookRepository.findById(id);
-
         Book book;
         if(result.isPresent()) {
             book = result.get();
             return(book);
+        }else{
+            // Will throw error if a book with that ID is not found
+            throw new BookNotFoundException();
         }
-        else{
-            throw new RuntimeException("Did not find book id - " + id);
+    }
+
+    @Override
+    public Book updateBook(Book book) throws RuntimeException{
+        try{
+            Optional <Book> result = this.bookRepository.findByCode(book.getCode());
+            if(result.isEmpty()){
+                // Will throw error if a book with that code is not found
+                throw new BookNotFoundException();
+            }
+            book.setId(result.get().getId());
+            return this.bookRepository.save(book);
+        } catch (BookNotFoundException e) {
+            throw e;
+        } catch (RuntimeException e){
+            // Will throw error if any of the fields are empty.
+            throw new NullFieldsException();
         }
     }
 
     @Override
     // Adds a book to the database
     public Book save(Book book) {
-        return (this.bookRepository.save(book));
+        try{
+            Optional<Book> s = this.bookRepository.findByCode(book.getCode());
+            if(s.isPresent()){
+                throw new BookAlreadyExistsException();
+            }else{
+                return (this.bookRepository.save(book));
+            }
+        }catch (BookAlreadyExistsException e){
+            // Will throw error if a book with that Code already exists
+            throw e;
+        }catch (RuntimeException e){
+            // Will throw error if any of the fields are NULL.
+            throw new NullFieldsException();
+        }
     }
 
     @Override
     // Deletes a book that corresponds to an id
-    public void deleteById(int id) {
-        this.bookRepository.deleteById(id);
+    public Book deleteById(int id) {
+        Optional<Book> b = this.bookRepository.findById(id);
+        if (b.isEmpty()) {
+            // Will throw error if a book with that ID does not exist
+            throw new BookNotFoundException();
+        } else {
+            this.bookRepository.deleteById(id);
+            return b.get();
+        }
     }
 }
